backend/requirements.txt
fastapi
uvicorn
requests
alpaca-py
ccxt
